==> R/06_analysis_2.qmd <==
---
title: "06_analisys_2"
format:
  html:
    embed-resources: true
editor: visual
---
## Loading libraries

```{r}
#| message: false
library("tidyverse")
```


## Loading dataset

```{r}
tsv_file <- "../data/data_augmented.tsv"
data_augmented <- read_tsv(tsv_file, show_col_types = FALSE)
```

## Plotting Pathways

```{r}
# Enhanced stacked bar chart with ggplot2
plot_kegg <- ggplot(data_augmented |> 
             filter(!is.na(Compound_Name) & !is.na(Pathway)) |> 
             dplyr::count(Compound_Name, Pathway) |> 
             group_by(Compound_Name) |> 
             mutate(Proportion = n / sum(n)), 
           aes(fill = Pathway, y = Proportion, x = Compound_Name)) +
      geom_bar(position = "fill", stat = "identity") +
      scale_fill_viridis_d() +
      theme_minimal() +
      theme(
        axis.text.x = element_text(angle = 65, vjust = 1, hjust=1, size = 8, color = "black"),
        axis.text.y = element_text(size = 10, color = "black"),
        axis.title = element_text(size = 10, face = "bold"),
        plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
        legend.text = element_text(size = 8),
        legend.title = element_text(size = 10),
        
        ) +
      guides(fill = guide_legend(override.aes = list(size = 2))) +
      labs(
        title = "Proportion of Pathways for Each Compound",
        x = "Compound",
        y = "Proportion",
        fill = "Pathway"
      )


# Save the plot as a PNG file in the 'results' folder
ggsave("../results/compound_pathways_plot.png", plot_kegg, width = 12, height = 8, units = "in", dpi = 300)
```

From this plot we can observe which are the most common used pathways to biodegrade each compound, we can observe that there are some pathways that are commonly used for all the compounds, in the next graph its seen more clearly.

## Plotting which are the most common patwhays

```{r}
# Remove duplicates based on Compound_Name and Pathway, then count the occurrences of each pathway
pathway_counts <- data_augmented |> 
  filter(!is.na(Pathway)) |> 
  distinct(Compound_Name, Pathway, .keep_all = TRUE) |>   # Remove duplicates
  dplyr::count(Pathway) |> 
  mutate(Pathway = fct_reorder(Pathway, n)) |> # Reorder factors for plotting
  arrange(desc(n))

# Create an enhanced barplot
pathway_barplot <- ggplot(pathway_counts, aes(x = Pathway, y = n, fill = n)) +
  geom_bar(stat = "identity") +
  scale_fill_gradient(low = "skyblue", high = "blue") + # Gradient color
  coord_flip() +  # Flips the axes to make long labels readable
  theme_minimal(base_size = 14) + # Larger base font size
  theme(
    axis.text.x = element_text(hjust = 1),
    axis.text.y = element_text(size = 10), # Smaller text size for y axis labels
    axis.title.x = element_blank(), # Remove the x axis title
    axis.title.y = element_text(size = 14, face = "bold"),
    legend.position = "none" # Hide the legend as the color represents frequency
  ) +
  labs(
    title = "Most Used Pathways",
    x = "",
    y = "Frequency"
  ) +
  guides(fill = guide_colourbar(title = "Count")) + # Add a colorbar guide
  scale_y_continuous(limits = c(0, 15)) # Set y-axis limits

ggsave("../results/most_used_pathways.png", pathway_barplot, width = 12, height = 8, units = "in", dpi = 300)
```

In this graphic we can see that the most used pathways are the chlorocyclohexane and chlorobenzene degradation and the benzoate degradation. In the future, this information can be used to find new organisms that are able to biodegrade this compounds, we only need to search more organisms that are involved in this pathways. Also, it means that all the compounds that we want to biodegrade are very rich in this compounds.
